+++ system/src/Grav/Common/Service/AccountsServiceProvider.php
@@ -10,7 +10,6 @@
 namespace Grav\Common\Service;
 
 use Grav\Common\Config\Config;
-use Grav\Common\Debugger;
 use Grav\Common\User\DataUser;
 use Grav\Common\User\FlexUser;
 use Grav\Common\User\User;
@@ -26,40 +25,54 @@ class AccountsServiceProvider implements ServiceProviderInterface
     public function register(Container $container)
     {
         $container['accounts'] = function (Container $container) {
-            $type = strtolower(defined('GRAV_USER_INSTANCE') ? GRAV_USER_INSTANCE : $container['config']->get('system.accounts.type', 'data'));
-            if ($type === 'flex') {
-                /** @var Debugger $debugger */
-                $debugger = $container['debugger'];
-                $debugger->addMessage('User Accounts: Flex Directory');
-                return $this->flexAccounts($container);
-            }
-
-            return $this->dataAccounts($container);
+            $type = $this->initialize($container);
+            return $type === 'flex' ? $this->flexAccounts($container) : $this->dataAccounts($container);
         };
 
-        $container['users'] = $container->factory(function (Container $container) {
+        $container['user_groups'] = static function () {
+            return (new FlexDirectory('grav-user-groups', 'blueprints://flex/user-groups.yaml', ['enabled' => true]))->getIndex();
+        };
+
+        $container['users'] = $container->factory(static function (Container $container) {
             user_error('Grav::instance()[\'users\'] is deprecated since Grav 1.6, use Grav::instance()[\'accounts\'] instead', E_USER_DEPRECATED);
 
             return $container['accounts'];
         });
     }
 
-    protected function dataAccounts(Container $container)
+    protected function initialize(Container $container): string
     {
-        if (!defined('GRAV_USER_INSTANCE')) {
+        $isDefined = defined('GRAV_USER_INSTANCE');
+        $type = strtolower($isDefined ? GRAV_USER_INSTANCE : $container['config']->get('system.accounts.type', 'data'));
+
+        if ($type === 'flex') {
+            if (!$isDefined) {
+                define('GRAV_USER_INSTANCE', 'FLEX');
+            }
+
+            /** @var EventDispatcher $dispatcher */
+            $dispatcher = $container['events'];
+            $dispatcher->addListener('onFlexInit', static function (Event $event) use ($container) {
+                /** @var Flex $flex */
+                $flex = $event['flex'];
+                $flex->addDirectory($container['accounts']->getFlexDirectory());
+                $flex->addDirectory($container['user_groups']->getFlexDirectory());
+            });
+        } elseif (!$isDefined) {
             define('GRAV_USER_INSTANCE', 'DATA');
         }
 
+        return $type;
+    }
+
+    protected function dataAccounts(Container $container)
+    {
         // Use User class for backwards compatibility.
         return new DataUser\UserCollection(User::class);
     }
 
     protected function flexAccounts(Container $container)
     {
-        if (!defined('GRAV_USER_INSTANCE')) {
-            define('GRAV_USER_INSTANCE', 'FLEX');
-        }
-
         /** @var Config $config */
         $config = $container['config'];
 
@@ -70,15 +83,7 @@ class AccountsServiceProvider implements ServiceProviderInterface
             ]
         ] + ($config->get('plugins.flex-objects.object') ?: []);
 
-        $directory = new FlexDirectory('accounts', 'blueprints://flex/accounts.yaml', $options);
-
-        /** @var EventDispatcher $dispatcher */
-        $dispatcher = $container['events'];
-        $dispatcher->addListener('onFlexInit', static function (Event $event) use ($directory) {
-            /** @var Flex $flex */
-            $flex = $event['flex'];
-            $flex->addDirectory($directory);
-        });
+        $directory = new FlexDirectory('grav-accounts', 'blueprints://flex/accounts.yaml', $options);
 
         return $directory->getIndex();
     }