+++ system/src/Grav/Common/User/FlexUser/GroupCollection.php
@@ -1,0 +1,37 @@
+<?php
+
+/**
+ * @package    Grav\Common\User
+ *
+ * @copyright  Copyright (C) 2015 - 2019 Trilby Media, LLC. All rights reserved.
+ * @license    MIT License; see LICENSE file for details.
+ */
+
+namespace Grav\Common\User\FlexUser;
+
+use Grav\Framework\Flex\FlexCollection;
+
+class GroupCollection extends FlexCollection
+{
+    /**
+     * Checks user authorization to the action.
+     *
+     * @param  string $action
+     * @param  string|null $scope
+     * @return bool|null
+     */
+    public function authorize(string $action, string $scope = null): ?bool
+    {
+        $authorized = null;
+        foreach ($this as $object) {
+            $auth = $object->authorize($action, $scope);
+            if ($auth === true) {
+                $authorized = true;
+            } elseif ($auth === false) {
+                return false;
+            }
+        }
+
+        return $authorized;
+    }
+}