+++ system/src/Grav/Common/User/FlexUser/GroupIndex.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @package    Grav\Common\User
+ *
+ * @copyright  Copyright (C) 2015 - 2019 Trilby Media, LLC. All rights reserved.
+ * @license    MIT License; see LICENSE file for details.
+ */
+
+namespace Grav\Common\User\FlexUser;
+
+use Grav\Framework\Flex\FlexIndex;
+
+class GroupIndex extends FlexIndex
+{
+}