+++ system/src/Grav/Common/User/FlexUser/Storage/UserFolderStorage.php
@@ -13,4 +13,19 @@ use Grav\Framework\Flex\Storage\FolderStorage;
 
 class UserFolderStorage extends FolderStorage
 {
+    /**
+     * Prepares the row for saving and returns the storage key for the record.
+     *
+     * @param array $row
+     */
+    protected function prepareRow(array &$row): void
+    {
+        parent::prepareRow($row);
+
+        $access = $row['access'] ?? [];
+        unset($row['access']);
+        if ($access) {
+            $row['access'] = $access;
+        }
+    }
 }